Adds convenience functions to onc.util
dt2str takes a python datetime object and converts it to an ISO8601 string that is compatible with the Oceans 3.0 API.

get_onc_token will grab a token that is placed in the password portion of a .netrc entry. Use of this function requires that users add that information to a .netrc file themselves.